Add files via upload
User chooses option 1, 2, or 3 (choosing 4 will exit the program).  Option one allows the user to input a weight, which will be calculated from 5 fake planets' gravity. Option two will add up the items in the code "cash register" and return a total.  Option 3 calls for the user to input a question and then returns one of 19 random answers from a Magic-8-Ball.  The program will continue to loop until the user enters a 4 and the program exits.